Add support for transformations
Add preliminary support for transformations. Transformations act on HTML
nodes and modify them. For example in this checkin we introduce the
transformation called trim. When applying the trim transformation on an
HTML note, it takes the text contents of the HTML node and trims the
whitespace to the left and right, then returns the resulting test.